requirements_enhanced.txt
ultralytics>=8.0.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
matplotlib>=3.7.0
segment-anything>=1.0
transformers>=4.30.0
salesforce-lavis @ git+https://github.com/salesforce/LAVIS.git
clip @ git+https://github.com/openai/CLIP.git
flask>=2.3.0
flask-cors>=4.0.0
tqdm>=4.65.0
pyyaml>=6.0
requests>=2.31.0
scikit-learn>=1.3.0
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
openai
dotenv